Still learning about backwards compatibility... again support v8 and v9

+++ installScripts/updatePrebuilt/u1-update-pilotAPI.sh
@@ -3,13 +3,18 @@
 # This will download the prebuilt package for pilotapi and copy the
 # files to the program directory. As it does not include config files,
 # it is not suited for initial installation
+# usage: sudo sh u1-update-pilotAPI.sh net9.0
 
 if [ `whoami` != root ]; then
     echo Please run this script using sudo
     exit
 fi
+if [ "$#" -ne 1 ]; then
+        echo please pass net version as parameter, e.g. sudo sh u1-update-pilotAPI.sh net9.0
+        exit
+fi
 echo downloading release
-wget https://roethenmund.biz/pilot/pilotapi.tar.gz
+wget https://roethenmund.biz/pilot/pilotapi_$1.tar.gz
 echo stopping services
 systemctl stop gpsLogger
 systemctl stop sensorsLogger
@@ -19,7 +24,7 @@ mkdir temp
 cp -r /opt/pilotapi/config temp
 cp -r /opt/pilotapi/PiLot.API.dll.config temp
 rm -r /opt/pilotapi/*
-tar zxf pilotapi.tar.gz -C /opt/pilotapi
+tar zxf pilotapi_$1.tar.gz -C /opt/pilotapi
 cp -r temp/* /opt/pilotapi/
 echo starting services
 systemctl start gpsLogger
@@ -27,5 +32,5 @@ systemctl start sensorsLogger
 systemctl start pilotApi
 echo cleaning up
 rm -r temp
-rm pilotapi.tar.gz
+rm pilotapi_$1.tar.gz
 echo done
