+++ installScripts/updatePrebuilt/u6-update-pilotSensors.sh
@@ -3,25 +3,29 @@
 # This will download the prebuilt package for PiLot Sensors Logger and copy
 # the files to the program directory. As it does not include config files,
 # it is not suited for initial installation
+# usage: sudo sh u6-update-pilotSensors.sh net9.0
 
 if [ `whoami` != root ]; then
     echo Please run this script using sudo
     exit
 fi
-
+if [ "$#" -ne 1 ]; then
+        echo please pass net version as parameter, e.g. sudo sh u6-update-pilotSensors.sh net9.0
+        exit
+fi
 echo downloading release
-wget https://roethenmund.biz/pilot/pilotsensors.tar.gz
+wget https://roethenmund.biz/pilot/pilotsensors_$1.tar.gz
 echo stopping services
 systemctl stop sensorsLogger
 echo installing application
 mkdir temp
 mv /opt/pilotsensors/PiLot.Sensors.dll.config temp
 rm -r /opt/pilotsensors/*
-tar zxf pilotsensors.tar.gz -C /opt/pilotsensors
+tar zxf pilotsensors_$1.tar.gz -C /opt/pilotsensors
 mv temp/* /opt/pilotsensors
 echo starting services
 systemctl start sensorsLogger
 echo cleaning up
 rm -r temp
-rm pilotsensors.tar.gz
+rm pilotsensors_$1.tar.gz
 echo done
